# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记账 API。部分业务错误使用 HTTP 200，通过 status/msg 判断结果。",
        "title": "Jiezhang API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/friends": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "协作者列表",
                "operationId": "FriendsHandler_List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.FriendListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/friends/accept_apply": {
            "post": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "接受邀请，Body: `{invite_token, nickname}`",
                "operationId": "FriendsHandler_AcceptApply",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FriendAcceptApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/friends/invite": {
            "post": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "邀请，Body: `{account_book_id, role}`",
                "operationId": "FriendsHandler_Invite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FriendInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/friends/invite_information": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "查看邀请信息，Query: `invite_token`",
                "operationId": "FriendsHandler_InviteInformation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.InviteInformationItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/friends/{collaboratorId}": {
            "put": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "更新协作者，Body: `{account_book_id, role?, remark?}`",
                "operationId": "FriendsHandler_Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "collaboratorId",
                        "name": "collaboratorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FriendUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "移除协作者",
                "operationId": "FriendsHandler_Remove",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "collaboratorId",
                        "name": "collaboratorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FriendRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.FriendAcceptApplyRequest": {
            "type": "object",
            "properties": {
                "invite_token": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "types.FriendAuthority": {
            "type": "object",
            "properties": {
                "change_role": {
                    "type": "boolean"
                },
                "remove": {
                    "type": "boolean"
                }
            }
        },
        "types.FriendCollaboratorItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/types.FriendUserItem"
                }
            }
        },
        "types.FriendInviteRequest": {
            "type": "object",
            "properties": {
                "account_book_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "types.FriendListResponse": {
            "type": "object",
            "properties": {
                "authority": {
                    "$ref": "#/definitions/types.FriendAuthority"
                },
                "collaborators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.FriendCollaboratorItem"
                    }
                },
                "owner": {
                    "$ref": "#/definitions/types.FriendUserItem"
                }
            }
        },
        "types.FriendRemoveRequest": {
            "type": "object",
            "properties": {
                "account_book_id": {
                    "type": "integer"
                }
            }
        },
        "types.FriendUpdateRequest": {
            "type": "object",
            "properties": {
                "account_book_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "types.FriendUserItem": {
            "type": "object",
            "properties": {
                "avatar_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "types.InviteInformationBook": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.InviteInformationItem": {
            "type": "object",
            "properties": {
                "account_book": {
                    "$ref": "#/definitions/types.InviteInformationBook"
                },
                "invite_user": {
                    "$ref": "#/definitions/types.FriendUserItem"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppID": {
            "type": "apiKey",
            "name": "X-WX-APP-ID",
            "in": "header"
        },
        "SessionKey": {
            "type": "apiKey",
            "name": "X-WX-Skey",
            "in": "header"
        }
    }
}